+++ ddocserver/internals/functions/general/eioupdateindicator.sql
@@ -1,4 +1,4 @@
-create or replace function eioUpdateIndicator(int4, int4, varchar, timestamp, timestamp, timestamp, int4, int4, varchar) returns int4 as
+create or replace function eioUpdateIndicator(int8, int4, varchar, timestamp, timestamp, timestamp, int4, int4, varchar) returns int4 as
 $BODY$
 declare
     idRec alias for $1;
@@ -40,7 +40,7 @@ end
 $BODY$
 language 'plpgsql';
 
-create or replace function eioCatIndExist(int4, int4) returns bool as
+create or replace function eioCatIndExist(int8, int4) returns bool as
 $BODY$
 declare
     idRec alias for $1;
